refactor(logging): 使用 lombok @Slf4j 替换显式 Logger 定义并统一日志调用
- 移除显式声明的 Logger 字段，改用 @Slf4j 注解自动生成日志实例
- 更新所有日志调用由 LOG.xxx 变更为 log.xxx，简化代码风格
- 统一日志导入，移除对 org.slf4j.Logger 和 LoggerFactory 的依赖
- ApiSignatureFilter 中调整 ContentCachingRequestWrapper 以支持多次读取请求体
- Message 组件模板缓存替换底层缓存实现为 ydsz-common-cache，优化缓存管理与指标采集
- 修正 CasController 中 HTML 文本拼接为 String.format，提高代码可读性
- System 模块添加接口权限注解，增强接口安全控制
- 保持代码风格一致性，增强维护性和扩展性

+++ ydsz-userinfo/ydsz-userinfo-server/src/main/java/com/njydsz/userinfo/server/auth/SecurityDashboardService.java
@@ -161,7 +161,7 @@ public class SecurityDashboardService {
 
     // 查询当日所有失败记录
     Map<String, Integer> reasonCountMap = new HashMap<>();
-    int totalFails = countLoginsByResult(dayStart, dayEnd, "FAILED");
+    int totalFails = (int) countLoginsByResult(dayStart, dayEnd, "FAILED");
 
     if (totalFails == 0) {
       return List.of();
